Remove deploy script
It appears that shell_exec() is not allowed for browser requests (it only returns null when run in the browser, but works normally when this file is run in PHP CLI). Instead of identifying and reverting this presumed security precaution, we'll try to set up a streamlined manual deployment process.